refactor: extract GraphQL error classes to src/gql/errors/
- Create src/gql/errors/ directory with separate error classes
- Add AuthenticationError, ForbiddenError, ValidationError
- Add index.js to export all error classes
- Update authValidations.js to import from new location
