+++ src/gql/errors/index.js
@@ -1,0 +1,11 @@
+'use strict';
+
+const AuthenticationError = require('./AuthenticationError');
+const ForbiddenError = require('./ForbiddenError');
+const ValidationError = require('./ValidationError');
+
+module.exports = {
+	AuthenticationError,
+	ForbiddenError,
+	ValidationError,
+};
